Revert "🔧 Refactor awards system: Extract components and fix architecture issues"
This reverts commit 9a110250947f3e05e0aa4b1b0cf6f162c9468392.
