+++ GitStreak/ContentView.swift
@@ -657,44 +657,535 @@ struct SettingsView: View {
     }
 }
 
-// MARK: - Temporary AwardsTabView (until files added to Xcode project)
+#Preview {
+    ContentView()
+}
+
+// MARK: - Awards Tab Implementation
+
 struct AwardsTabView: View {
     @ObservedObject var dataModel: GitStreakDataModel
+    @State private var expandedCategories: Set<AchievementCategory> = []
+    
+    private var achievementsByCategory: [AchievementCategory: [Achievement]] {
+        Dictionary(grouping: dataModel.achievements, by: { $0.category })
+    }
     
     var body: some View {
-        VStack(spacing: 20) {
-            Text("Awards")
-                .font(.largeTitle)
-                .fontWeight(.bold)
-            
-            Text("Complete awards system with intelligent filtering")
-                .font(.subheadline)
-                .foregroundColor(.secondary)
-                .multilineTextAlignment(.center)
-            
-            Text("\(dataModel.unlockedAchievementCount)/\(dataModel.achievements.count) achievements unlocked")
-                .font(.title2)
-                .fontWeight(.semibold)
-                .foregroundColor(.green)
-            
-            Text("Total XP: \(dataModel.totalAchievementXP)")
-                .font(.title3)
-                .foregroundColor(.blue)
-            
-            if !dataModel.recentlyUnlockedAchievements.isEmpty {
-                Text("Recent: \(dataModel.recentlyUnlockedAchievements.count) unlocked")
-                    .font(.subheadline)
-                    .foregroundColor(.orange)
+        ScrollView {
+            VStack(alignment: .leading, spacing: 24) {
+                // Summary Stats
+                AchievementSummaryView(dataModel: dataModel)
+                
+                // Achievement Categories
+                VStack(spacing: 16) {
+                    ForEach(AchievementCategory.allCases, id: \.self) { category in
+                        AchievementCategoryView(
+                            category: category,
+                            achievements: achievementsByCategory[category] ?? [],
+                            isExpanded: expandedCategories.contains(category),
+                            onToggle: {
+                                if expandedCategories.contains(category) {
+                                    expandedCategories.remove(category)
+                                } else {
+                                    expandedCategories.insert(category)
+                                }
+                            },
+                            dataModel: dataModel
+                        )
+                    }
+                }
+                
+                // Recent Achievements
+                if !dataModel.recentlyUnlockedAchievements.isEmpty {
+                    VStack(alignment: .leading, spacing: 16) {
+                        Text("Recent Achievements")
+                            .font(.system(size: 18, weight: .semibold))
+                            .foregroundColor(.primary)
+                        
+                        ScrollView(.horizontal, showsIndicators: false) {
+                            HStack(spacing: 12) {
+                                ForEach(dataModel.recentlyUnlockedAchievements.prefix(5)) { achievement in
+                                    CompactAchievementCardView(achievement: achievement)
+                                }
+                            }
+                            .padding(.horizontal, 24)
+                        }
+                    }
+                }
             }
-            
-            Spacer()
+            .padding(.horizontal, 24)
+            .padding(.vertical, 20)
         }
-        .padding()
         .navigationTitle("Awards")
         .background(Color(.systemGroupedBackground))
     }
 }
 
-#Preview {
-    ContentView()
+struct AchievementSummaryView: View {
+    @ObservedObject var dataModel: GitStreakDataModel
+    
+    private var totalAchievements: Int {
+        dataModel.achievements.count
+    }
+    
+    private var nextMilestone: String {
+        let unlocked = dataModel.unlockedAchievementCount
+        let nextTargets = [10, 25, 50, 75, 100]
+        
+        for target in nextTargets {
+            if unlocked < target {
+                return "Achievement Hunter (\(target) total)"
+            }
+        }
+        return "All achievements unlocked!"
+    }
+    
+    private var milestoneProgress: Double {
+        let unlocked = dataModel.unlockedAchievementCount
+        let nextTargets = [10, 25, 50, 75, 100]
+        
+        for target in nextTargets {
+            if unlocked < target {
+                return Double(unlocked) / Double(target)
+            }
+        }
+        return 1.0
+    }
+    
+    var body: some View {
+        VStack(spacing: 16) {
+            Text("Achievement Progress")
+                .font(.system(size: 18, weight: .semibold))
+                .foregroundColor(.primary)
+                .frame(maxWidth: .infinity, alignment: .leading)
+            
+            HStack(spacing: 16) {
+                // Total Achievements
+                StatCardView(
+                    title: "Achievements",
+                    value: "\(dataModel.unlockedAchievementCount)/\(totalAchievements)",
+                    color: .green
+                )
+                
+                // Achievement XP
+                StatCardView(
+                    title: "Bonus XP",
+                    value: "\(dataModel.totalAchievementXP.formatted())",
+                    color: .blue
+                )
+            }
+            
+            // Next Milestone Progress
+            VStack(alignment: .leading, spacing: 8) {
+                HStack {
+                    Text("Next Milestone")
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(.secondary)
+                    
+                    Spacer()
+                    
+                    Text(nextMilestone)
+                        .font(.system(size: 12, weight: .medium))
+                        .foregroundColor(.primary)
+                }
+                
+                ProgressBarView(progress: milestoneProgress, height: 6)
+            }
+        }
+        .padding(20)
+        .background(Color(.systemBackground))
+        .cornerRadius(16)
+        .shadow(color: .black.opacity(0.1), radius: 5, x: 0, y: 2)
+    }
+}
+
+struct StatCardView: View {
+    let title: String
+    let value: String
+    let color: Color
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 4) {
+            Text(title)
+                .font(.system(size: 12, weight: .medium))
+                .foregroundColor(.secondary)
+            
+            Text(value)
+                .font(.system(size: 16, weight: .semibold))
+                .foregroundColor(color)
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .padding(12)
+        .background(color.opacity(0.1))
+        .cornerRadius(12)
+    }
+}
+
+struct AchievementCategoryView: View {
+    let category: AchievementCategory
+    let achievements: [Achievement]
+    let isExpanded: Bool
+    let onToggle: () -> Void
+    @ObservedObject var dataModel: GitStreakDataModel
+    
+    private var unlockedCount: Int {
+        achievements.filter { $0.unlocked }.count
+    }
+    
+    private var progressPercentage: Double {
+        guard !achievements.isEmpty else { return 0 }
+        return Double(unlockedCount) / Double(achievements.count)
+    }
+    
+    private func getVisibleAchievements(for categoryAchievements: [Achievement]) -> [Achievement] {
+        var visibleAchievements: [Achievement] = []
+        
+        // Show unlocked achievements first (limit to 3 max)
+        let unlockedAchievements = categoryAchievements.filter { $0.unlocked }
+        let displayedUnlocked = Array(unlockedAchievements.prefix(3))
+        visibleAchievements.append(contentsOf: displayedUnlocked)
+        
+        // Calculate remaining slots (max 3 total)
+        let remainingSlots = max(0, 3 - displayedUnlocked.count)
+        
+        // Show next achievable ones based on remaining slots
+        if remainingSlots > 0 {
+            let lockedAchievements = categoryAchievements.filter { !$0.unlocked }
+            let nextAchievements = getNextAchievableAchievements(from: lockedAchievements, category: self.category, maxCount: remainingSlots)
+            visibleAchievements.append(contentsOf: nextAchievements)
+        }
+        
+        return visibleAchievements
+    }
+    
+    private func getNextAchievableAchievements(from lockedAchievements: [Achievement], category: AchievementCategory, maxCount: Int) -> [Achievement] {
+        // Get current user stats to determine next logical achievements
+        let currentStreak = dataModel.currentStreak
+        let weeklyCommits = dataModel.totalCommitsThisWeek
+        let currentXP = dataModel.xp
+        
+        // Sort locked achievements by difficulty/requirement to show most achievable first
+        let sortedAchievements = lockedAchievements.sorted { first, second in
+            // Sort by XP value (assuming lower XP = easier/next logical step)
+            getAchievementXP(for: first.title) < getAchievementXP(for: second.title)
+        }
+        
+        // Determine ideal count based on category and user progress, but respect maxCount limit
+        let idealCount: Int
+        switch category {
+        case .streaks:
+            idealCount = currentStreak == 0 ? 1 : 2
+            
+        case .volume:
+            idealCount = weeklyCommits < 10 ? 1 : 2
+            
+        case .dailyPatterns:
+            idealCount = 3
+            
+        case .weeklyPatterns:
+            idealCount = 2
+            
+        case .codeImpact:
+            idealCount = currentXP < 500 ? 1 : 2
+            
+        case .repositoryDiversity:
+            idealCount = 2
+            
+        case .specialMilestones:
+            idealCount = currentXP < 1000 ? 1 : 2
+        }
+        
+        let finalCount = min(idealCount, maxCount)
+        return Array(sortedAchievements.prefix(finalCount))
+    }
+    
+    private func getAchievementXP(for title: String) -> Int {
+        // Map achievement titles to their XP values for sorting
+        switch title {
+        // Streak achievements (25-500 XP)
+        case "First Flame": return 25
+        case "Getting Warmed Up": return 50
+        case "Week Warrior": return 100
+        case "Fortnight Fighter": return 200
+        case "Monthly Master": return 350
+        case "Quarter Champion": return 500
+        case "Semester Superstar": return 750
+        case "Annual Achiever": return 1000
+        case "Eternal Flame": return 2000
+        
+        // Volume achievements (25-400 XP)
+        case "First Steps": return 25
+        case "Getting Started": return 50
+        case "Century Club": return 100
+        case "Volume Champion": return 200
+        case "Commit Crusher": return 300
+        case "Thousand Club": return 400
+        
+        // Daily pattern achievements (75-350 XP)
+        case "Early Bird": return 75
+        case "Night Owl": return 75
+        case "Consistent Coder": return 150
+        case "Daily Devotion": return 200
+        case "Perfect Week": return 300
+        case "Unstoppable": return 350
+        
+        // Weekly pattern achievements (100-400 XP)
+        case "Weekend Warrior": return 100
+        case "Weekday Hero": return 100
+        case "Balanced Builder": return 200
+        case "Steady Pace": return 300
+        case "Weekly Wonder": return 400
+        
+        // Code impact achievements (100-600 XP)
+        case "File Modifier": return 100
+        case "Multi-File Master": return 200
+        case "Repository Ruler": return 300
+        case "Code Architect": return 400
+        case "Impact Legend": return 500
+        case "Change Champion": return 600
+        
+        // Repository diversity achievements (150-500 XP)
+        case "Explorer": return 150
+        case "Multi-Talented": return 250
+        case "Language Learner": return 300
+        case "Polyglot Programmer": return 400
+        case "Diversity Champion": return 500
+        
+        // Special milestones (200-1500 XP)
+        case "First Year": return 200
+        case "Dedication Award": return 400
+        case "Consistency King": return 600
+        case "Milestone Master": return 800
+        case "Achievement Hunter (10 total)": return 300
+        case "Achievement Hunter (25 total)": return 600
+        case "Achievement Hunter (50 total)": return 1000
+        case "Achievement Hunter (75 total)": return 1200
+        case "Ultimate Achiever": return 1500
+        
+        default: return 999 // Unknown achievements get high value (shown last)
+        }
+    }
+    
+    var body: some View {
+        VStack(spacing: 0) {
+            // Category Header
+            Button(action: onToggle) {
+                HStack(spacing: 12) {
+                    // Category Icon with Progress Circle
+                    ZStack {
+                        Circle()
+                            .stroke(Color.gray.opacity(0.3), lineWidth: 3)
+                            .frame(width: 40, height: 40)
+                        
+                        Circle()
+                            .trim(from: 0, to: progressPercentage)
+                            .stroke(
+                                LinearGradient(
+                                    gradient: Gradient(colors: [
+                                        Color(red: 0.2, green: 0.8, blue: 0.4),
+                                        Color(red: 0.3, green: 0.5, blue: 0.9),
+                                        Color(red: 0.6, green: 0.3, blue: 0.9)
+                                    ]),
+                                    startPoint: .leading,
+                                    endPoint: .trailing
+                                ),
+                                style: StrokeStyle(lineWidth: 3, lineCap: .round)
+                            )
+                            .rotationEffect(.degrees(-90))
+                            .frame(width: 40, height: 40)
+                        
+                        Text(category.emoji)
+                            .font(.title3)
+                    }
+                    
+                    VStack(alignment: .leading, spacing: 2) {
+                        Text(category.displayName)
+                            .font(.system(size: 16, weight: .semibold))
+                            .foregroundColor(.primary)
+                        
+                        Text("\(unlockedCount)/\(achievements.count) unlocked")
+                            .font(.system(size: 14, weight: .medium))
+                            .foregroundColor(.secondary)
+                    }
+                    
+                    Spacer()
+                    
+                    Image(systemName: isExpanded ? "chevron.up" : "chevron.down")
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(.secondary)
+                }
+                .padding(16)
+                .background(Color(.systemBackground))
+                .cornerRadius(12)
+            }
+            .buttonStyle(PlainButtonStyle())
+            
+            // Achievement Cards (when expanded) - Only show unlocked + next achievable
+            if isExpanded {
+                VStack(spacing: 8) {
+                    ForEach(getVisibleAchievements(for: achievements)) { achievement in
+                        SimpleAchievementCardView(achievement: achievement)
+                    }
+                }
+                .padding(.top, 8)
+            }
+        }
+    }
+}
+
+struct SimpleAchievementCardView: View {
+    let achievement: Achievement
+    
+    var body: some View {
+        HStack(spacing: 12) {
+            // Achievement Icon
+            Text(achievement.icon)
+                .font(.title2)
+                .frame(width: 32, height: 32)
+            
+            VStack(alignment: .leading, spacing: 4) {
+                Text(achievement.title)
+                    .font(.system(size: 15, weight: .semibold))
+                    .foregroundColor(achievement.unlocked ? .primary : .secondary)
+                
+                Text(achievement.description)
+                    .font(.system(size: 13, weight: .regular))
+                    .foregroundColor(.secondary)
+                    .multilineTextAlignment(.leading)
+            }
+            
+            Spacer()
+            
+            VStack(alignment: .trailing, spacing: 4) {
+                // Status Badge
+                if achievement.unlocked {
+                    HStack(spacing: 4) {
+                        Image(systemName: "checkmark")
+                            .font(.system(size: 10, weight: .bold))
+                        Text("UNLOCKED")
+                            .font(.system(size: 10, weight: .bold))
+                    }
+                    .foregroundColor(.green)
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 2)
+                    .background(Color.green.opacity(0.1))
+                    .cornerRadius(6)
+                } else {
+                    HStack(spacing: 4) {
+                        Image(systemName: "lock")
+                            .font(.system(size: 10, weight: .bold))
+                        Text("LOCKED")
+                            .font(.system(size: 10, weight: .bold))
+                    }
+                    .foregroundColor(.gray)
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 2)
+                    .background(Color.gray.opacity(0.1))
+                    .cornerRadius(6)
+                }
+            }
+        }
+        .padding(12)
+        .background(achievement.unlocked ? Color(.systemBackground) : Color.gray.opacity(0.05))
+        .overlay(
+            RoundedRectangle(cornerRadius: 12)
+                .stroke(
+                    achievement.unlocked ?
+                    LinearGradient(
+                        gradient: Gradient(colors: [
+                            Color(red: 0.2, green: 0.8, blue: 0.4),
+                            Color(red: 0.3, green: 0.5, blue: 0.9),
+                            Color(red: 0.6, green: 0.3, blue: 0.9)
+                        ]),
+                        startPoint: .leading,
+                        endPoint: .trailing
+                    ) :
+                    LinearGradient(colors: [Color.clear], startPoint: .leading, endPoint: .trailing),
+                    lineWidth: achievement.unlocked ? 2 : 0
+                )
+        )
+        .cornerRadius(12)
+        .shadow(color: .black.opacity(0.1), radius: 5, x: 0, y: 2)
+    }
+}
+
+struct ProgressBarView: View {
+    let progress: Double
+    let height: CGFloat
+    let cornerRadius: CGFloat
+    
+    init(progress: Double, height: CGFloat = 8, cornerRadius: CGFloat? = nil) {
+        self.progress = progress
+        self.height = height
+        self.cornerRadius = cornerRadius ?? height / 2
+    }
+    
+    var body: some View {
+        GeometryReader { geometry in
+            ZStack(alignment: .leading) {
+                RoundedRectangle(cornerRadius: cornerRadius)
+                    .fill(Color(.systemGray5))
+                    .frame(height: height)
+                
+                RoundedRectangle(cornerRadius: cornerRadius)
+                    .fill(
+                        LinearGradient(
+                            gradient: Gradient(colors: [
+                                Color(red: 0.2, green: 0.8, blue: 0.4),
+                                Color(red: 0.3, green: 0.5, blue: 0.9),
+                                Color(red: 0.6, green: 0.3, blue: 0.9)
+                            ]),
+                            startPoint: .leading,
+                            endPoint: .trailing
+                        )
+                    )
+                    .frame(width: geometry.size.width * progress, height: height)
+            }
+        }
+        .frame(height: height)
+    }
+}
+
+struct CompactAchievementCardView: View {
+    let achievement: Achievement
+    
+    var body: some View {
+        VStack(spacing: 8) {
+            Text(achievement.icon)
+                .font(.title)
+            
+            VStack(spacing: 2) {
+                Text(achievement.title)
+                    .font(.system(size: 12, weight: .semibold))
+                    .foregroundColor(.primary)
+                    .multilineTextAlignment(.center)
+                    .lineLimit(2)
+                
+                Text("Unlocked")
+                    .font(.system(size: 10, weight: .medium))
+                    .foregroundColor(.green)
+            }
+        }
+        .frame(width: 100, height: 80)
+        .padding(8)
+        .background(Color(.systemBackground))
+        .overlay(
+            RoundedRectangle(cornerRadius: 12)
+                .stroke(
+                    LinearGradient(
+                        gradient: Gradient(colors: [
+                            Color(red: 0.2, green: 0.8, blue: 0.4),
+                            Color(red: 0.3, green: 0.5, blue: 0.9),
+                            Color(red: 0.6, green: 0.3, blue: 0.9)
+                        ]),
+                        startPoint: .leading,
+                        endPoint: .trailing
+                    ),
+                    lineWidth: 2
+                )
+        )
+        .cornerRadius(12)
+        .shadow(color: .black.opacity(0.1), radius: 5, x: 0, y: 2)
+    }
 }
