+++ GitStreak/Views/AwardsTabView.swift
@@ -20,16 +20,14 @@ struct AwardsTabView: View {
                         AchievementCategoryView(
                             category: category,
                             achievements: achievementsByCategory[category] ?? [],
-                            isExpanded: expandedCategories.contains(category),
-                            onToggle: {
-                                if expandedCategories.contains(category) {
-                                    expandedCategories.remove(category)
-                                } else {
-                                    expandedCategories.insert(category)
-                                }
-                            },
-                            dataModel: dataModel
-                        )
+                            isExpanded: expandedCategories.contains(category)
+                        ) {
+                            if expandedCategories.contains(category) {
+                                expandedCategories.remove(category)
+                            } else {
+                                expandedCategories.insert(category)
+                            }
+                        }
                     }
                 }
                 
@@ -56,6 +54,302 @@ struct AwardsTabView: View {
         }
         .navigationTitle("Awards")
         .background(Color(.systemGroupedBackground))
+    }
+}
+
+struct AchievementSummaryView: View {
+    @ObservedObject var dataModel: GitStreakDataModel
+    
+    private var totalAchievements: Int {
+        dataModel.achievements.count
+    }
+    
+    private var nextMilestone: String {
+        let unlocked = dataModel.unlockedAchievementCount
+        let nextTargets = [10, 25, 50, 75, 100]
+        
+        for target in nextTargets {
+            if unlocked < target {
+                return "Achievement Hunter (\(target) total)"
+            }
+        }
+        return "All achievements unlocked!"
+    }
+    
+    private var milestoneProgress: Double {
+        let unlocked = dataModel.unlockedAchievementCount
+        let nextTargets = [10, 25, 50, 75, 100]
+        
+        for target in nextTargets {
+            if unlocked < target {
+                return Double(unlocked) / Double(target)
+            }
+        }
+        return 1.0
+    }
+    
+    var body: some View {
+        VStack(spacing: 16) {
+            Text("Achievement Progress")
+                .font(.system(size: 18, weight: .semibold))
+                .foregroundColor(.primary)
+                .frame(maxWidth: .infinity, alignment: .leading)
+            
+            HStack(spacing: 16) {
+                // Total Achievements
+                StatCardView(
+                    title: "Achievements",
+                    value: "\(dataModel.unlockedAchievementCount)/\(totalAchievements)",
+                    color: .green
+                )
+                
+                // Achievement XP
+                StatCardView(
+                    title: "Bonus XP",
+                    value: "\(dataModel.totalAchievementXP.formatted())",
+                    color: .blue
+                )
+            }
+            
+            // Next Milestone Progress
+            VStack(alignment: .leading, spacing: 8) {
+                HStack {
+                    Text("Next Milestone")
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(.secondary)
+                    
+                    Spacer()
+                    
+                    Text(nextMilestone)
+                        .font(.system(size: 12, weight: .medium))
+                        .foregroundColor(.primary)
+                }
+                
+                ProgressBarView(progress: milestoneProgress, height: 6)
+            }
+        }
+        .padding(20)
+        .background(Color(.systemBackground))
+        .cornerRadius(16)
+        .shadow(color: .black.opacity(0.1), radius: 5, x: 0, y: 2)
+    }
+}
+
+struct StatCardView: View {
+    let title: String
+    let value: String
+    let color: Color
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 4) {
+            Text(title)
+                .font(.system(size: 12, weight: .medium))
+                .foregroundColor(.secondary)
+            
+            Text(value)
+                .font(.system(size: 16, weight: .semibold))
+                .foregroundColor(color)
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .padding(12)
+        .background(color.opacity(0.1))
+        .cornerRadius(12)
+    }
+}
+
+struct AchievementCategoryView: View {
+    let category: AchievementCategory
+    let achievements: [Achievement]
+    let isExpanded: Bool
+    let onToggle: () -> Void
+    
+    private var unlockedCount: Int {
+        achievements.filter { $0.unlocked }.count
+    }
+    
+    private var progressPercentage: Double {
+        guard !achievements.isEmpty else { return 0 }
+        return Double(unlockedCount) / Double(achievements.count)
+    }
+    
+    var body: some View {
+        VStack(spacing: 0) {
+            // Category Header
+            Button(action: onToggle) {
+                HStack(spacing: 12) {
+                    // Category Icon with Progress Circle
+                    ZStack {
+                        Circle()
+                            .stroke(Color.gray.opacity(0.3), lineWidth: 3)
+                            .frame(width: 40, height: 40)
+                        
+                        Circle()
+                            .trim(from: 0, to: progressPercentage)
+                            .stroke(
+                                LinearGradient(
+                                    gradient: Gradient(colors: [
+                                        Color(red: 0.2, green: 0.8, blue: 0.4),
+                                        Color(red: 0.3, green: 0.5, blue: 0.9),
+                                        Color(red: 0.6, green: 0.3, blue: 0.9)
+                                    ]),
+                                    startPoint: .leading,
+                                    endPoint: .trailing
+                                ),
+                                style: StrokeStyle(lineWidth: 3, lineCap: .round)
+                            )
+                            .rotationEffect(.degrees(-90))
+                            .frame(width: 40, height: 40)
+                        
+                        Text(category.emoji)
+                            .font(.title3)
+                    }
+                    
+                    VStack(alignment: .leading, spacing: 2) {
+                        Text(category.displayName)
+                            .font(.system(size: 16, weight: .semibold))
+                            .foregroundColor(.primary)
+                        
+                        Text("\(unlockedCount)/\(achievements.count) unlocked")
+                            .font(.system(size: 14, weight: .medium))
+                            .foregroundColor(.secondary)
+                    }
+                    
+                    Spacer()
+                    
+                    Image(systemName: isExpanded ? "chevron.up" : "chevron.down")
+                        .font(.system(size: 14, weight: .medium))
+                        .foregroundColor(.secondary)
+                }
+                .padding(16)
+                .background(Color(.systemBackground))
+                .cornerRadius(12)
+            }
+            .buttonStyle(PlainButtonStyle())
+            
+            // Achievement Cards (when expanded)
+            if isExpanded {
+                VStack(spacing: 8) {
+                    ForEach(achievements) { achievement in
+                        SimpleAchievementCardView(achievement: achievement)
+                    }
+                }
+                .padding(.top, 8)
+            }
+        }
+    }
+}
+
+// Simple achievement card that doesn't conflict with existing AchievementCardView
+struct SimpleAchievementCardView: View {
+    let achievement: Achievement
+    
+    var body: some View {
+        HStack(spacing: 12) {
+            // Achievement Icon
+            Text(achievement.icon)
+                .font(.title2)
+                .frame(width: 32, height: 32)
+            
+            VStack(alignment: .leading, spacing: 4) {
+                Text(achievement.title)
+                    .font(.system(size: 15, weight: .semibold))
+                    .foregroundColor(achievement.unlocked ? .primary : .secondary)
+                
+                Text(achievement.description)
+                    .font(.system(size: 13, weight: .regular))
+                    .foregroundColor(.secondary)
+                    .multilineTextAlignment(.leading)
+            }
+            
+            Spacer()
+            
+            VStack(alignment: .trailing, spacing: 4) {
+                // Status Badge
+                if achievement.unlocked {
+                    HStack(spacing: 4) {
+                        Image(systemName: "checkmark")
+                            .font(.system(size: 10, weight: .bold))
+                        Text("UNLOCKED")
+                            .font(.system(size: 10, weight: .bold))
+                    }
+                    .foregroundColor(.green)
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 2)
+                    .background(Color.green.opacity(0.1))
+                    .cornerRadius(6)
+                } else {
+                    HStack(spacing: 4) {
+                        Image(systemName: "lock")
+                            .font(.system(size: 10, weight: .bold))
+                        Text("LOCKED")
+                            .font(.system(size: 10, weight: .bold))
+                    }
+                    .foregroundColor(.gray)
+                    .padding(.horizontal, 6)
+                    .padding(.vertical, 2)
+                    .background(Color.gray.opacity(0.1))
+                    .cornerRadius(6)
+                }
+            }
+        }
+        .padding(12)
+        .background(achievement.unlocked ? Color(.systemBackground) : Color.gray.opacity(0.05))
+        .overlay(
+            RoundedRectangle(cornerRadius: 12)
+                .stroke(
+                    achievement.unlocked ?
+                    LinearGradient(
+                        gradient: Gradient(colors: [
+                            Color(red: 0.2, green: 0.8, blue: 0.4),
+                            Color(red: 0.3, green: 0.5, blue: 0.9),
+                            Color(red: 0.6, green: 0.3, blue: 0.9)
+                        ]),
+                        startPoint: .leading,
+                        endPoint: .trailing
+                    ) :
+                    LinearGradient(colors: [Color.clear], startPoint: .leading, endPoint: .trailing),
+                    lineWidth: achievement.unlocked ? 2 : 0
+                )
+        )
+        .cornerRadius(12)
+        .shadow(color: .black.opacity(0.1), radius: 5, x: 0, y: 2)
+    }
+}
+
+struct ProgressBarView: View {
+    let progress: Double
+    let height: CGFloat
+    let cornerRadius: CGFloat
+    
+    init(progress: Double, height: CGFloat = 8, cornerRadius: CGFloat? = nil) {
+        self.progress = progress
+        self.height = height
+        self.cornerRadius = cornerRadius ?? height / 2
+    }
+    
+    var body: some View {
+        GeometryReader { geometry in
+            ZStack(alignment: .leading) {
+                RoundedRectangle(cornerRadius: cornerRadius)
+                    .fill(Color(.systemGray5))
+                    .frame(height: height)
+                
+                RoundedRectangle(cornerRadius: cornerRadius)
+                    .fill(
+                        LinearGradient(
+                            gradient: Gradient(colors: [
+                                Color(red: 0.2, green: 0.8, blue: 0.4),
+                                Color(red: 0.3, green: 0.5, blue: 0.9),
+                                Color(red: 0.6, green: 0.3, blue: 0.9)
+                            ]),
+                            startPoint: .leading,
+                            endPoint: .trailing
+                        )
+                    )
+                    .frame(width: geometry.size.width * progress, height: height)
+            }
+        }
+        .frame(height: height)
     }
 }
